refactor: 불필요한 인터페이스 추상화 제거
